Improve consensus sim framework:
Use tagged integers for validations, sequence and tx IDs.
Use generic Validations in Peer.
Flyweight style Ledger instances.

+++ src/ripple/consensus/Consensus.h
@@ -874,7 +874,7 @@ Consensus<Adaptor>::getJson(bool full) const
     if (mode_.get() != ConsensusMode::wrongLedger)
     {
         ret["synched"] = true;
-        ret["ledger_seq"] = previousLedger_.seq() + 1;
+        ret["ledger_seq"] = static_cast<std::uint32_t>(previousLedger_.seq() + 1);
         ret["close_granularity"] = static_cast<Int>(closeResolution_.count());
     }
     else
